Ensure prop protections are read/enforced in order
Property protections need to be enforced in the order specified by the
configuration file. i.e. [^x_owner.*] should be enforced before [.*]
given the example config.

To achieve this, use a list of tuples to store the rules rather than a
dict, and also ensure to specify dict_type when parsing the
configuration file for python 2.6.

Fixes bug 1224449

Change-Id: Ib22cb34b7daef081c95717b55a991b8d613468c4